Refactor: 에러메시지 함수, 클래스 생성 함수 모듈화
