+++ js/signup.js
@@ -1,22 +1,19 @@
 import { email, password, passwordCheck, passwordCheckVisible, passwordVisible, signupButton } from "./tags.js";
+import { signup, enterSignup, togglePasswordVisible, togglePasswordCheckVisible } from "./functions.js";
 import {
   addEmailErrorMsg,
   deleteEmailErrorMsg,
   addPassWordErrorMsgSignup,
   deletePasswordErrorMsg,
-  togglePasswordVisible,
-  checkPassword,
+  addPasswordCheckErrorMsg,
   deletePasswordCheckErrorMsg,
-  signup,
-  enterSignup,
-  togglePasswordCheckVisible,
-} from "./functions.js";
+} from "./errormsg.js";
 
 email.addEventListener("focusout", addEmailErrorMsg);
 email.addEventListener("focusin", deleteEmailErrorMsg);
 password.addEventListener("focusout", addPassWordErrorMsgSignup);
 password.addEventListener("focusin", deletePasswordErrorMsg);
-passwordCheck.addEventListener("focusout", checkPassword);
+passwordCheck.addEventListener("focusout", addPasswordCheckErrorMsg);
 passwordCheck.addEventListener("focusin", deletePasswordCheckErrorMsg);
 passwordVisible.addEventListener("click", togglePasswordVisible);
 passwordCheckVisible.addEventListener("click", togglePasswordCheckVisible);
